+++ tests/test_ecommerce.py
@@ -6,7 +6,7 @@ import scrapy
 from pydantic import ValidationError
 from scrapy_poet import DummyResponse
 from scrapy_spider_metadata import get_spider_metadata
-from zyte_common_items import Product, ProductNavigation
+from zyte_common_items import ProbabilityRequest, Product, ProductNavigation
 
 from zyte_spider_templates import BaseSpiderParams
 from zyte_spider_templates._geolocations import (
@@ -60,15 +60,15 @@ def test_crawl():
 
     subcategories = {
         "subCategories": [
-            {"url": subcategory_urls[0]},
-            {"url": subcategory_urls[1]},
+            {"url": subcategory_urls[0], "metadata": {"probability": 0.95}},
+            {"url": subcategory_urls[1], "metadata": {"probability": 0.78}},
         ],
     }
     nextpage = {"nextPage": {"url": nextpage_url}}
     items = {
         "items": [
-            {"url": item_urls[0]},
-            {"url": item_urls[1]},
+            {"url": item_urls[0], "metadata": {"probability": 0.99}},
+            {"url": item_urls[1], "metadata": {"probability": 0.83}},
         ],
     }
 
@@ -86,8 +86,10 @@ def test_crawl():
     assert len(requests) == 2
     assert requests[0].url == subcategory_urls[0]
     assert requests[0].callback == spider.parse_navigation
+    assert requests[0].priority == 95
     assert requests[1].url == subcategory_urls[1]
     assert requests[1].callback == spider.parse_navigation
+    assert requests[1].priority == 78
 
     # subcategories + nextpage
     navigation = ProductNavigation.from_dict(
@@ -102,6 +104,7 @@ def test_crawl():
     urls = {request.url for request in requests}
     assert urls == {*subcategory_urls, nextpage_url}
     assert all(request.callback == spider.parse_navigation for request in requests)
+    assert [request.priority for request in requests] == [100, 95, 78]
 
     # subcategories + nextpage + items
     navigation = ProductNavigation.from_dict(
@@ -120,6 +123,7 @@ def test_crawl():
             assert request.callback == spider.parse_product
         else:
             assert request.callback == spider.parse_navigation
+    assert [request.priority for request in requests] == [199, 183, 100, 95, 78]
 
     # nextpage + items
     navigation = ProductNavigation.from_dict(
@@ -137,6 +141,7 @@ def test_crawl():
     assert requests[1].callback == spider.parse_product
     assert requests[2].url == nextpage_url
     assert requests[2].callback == spider.parse_navigation
+    assert [request.priority for request in requests] == [199, 183, 100]
 
     # subcategories + items
     navigation = ProductNavigation.from_dict(
@@ -156,6 +161,7 @@ def test_crawl():
     assert requests[2].callback == spider.parse_navigation
     assert requests[3].url == subcategory_urls[1]
     assert requests[3].callback == spider.parse_navigation
+    assert [request.priority for request in requests] == [199, 183, 95, 78]
 
     # nextpage
     navigation = ProductNavigation.from_dict(
@@ -168,6 +174,7 @@ def test_crawl():
     assert len(requests) == 1
     assert requests[0].url == nextpage_url
     assert requests[0].callback == spider.parse_navigation
+    assert [request.priority for request in requests] == [100]
 
     # items
     navigation = ProductNavigation.from_dict(
@@ -182,6 +189,30 @@ def test_crawl():
     assert requests[0].callback == spider.parse_product
     assert requests[1].url == item_urls[1]
     assert requests[1].callback == spider.parse_product
+    assert [request.priority for request in requests] == [199, 183]
+
+    # Test parse_navigation() behavior on pagination_only crawl strategy.
+    spider = EcommerceSpider(
+        url="https://example.com/", crawl_strategy="pagination_only"
+    )
+
+    # nextpage + items
+    navigation = ProductNavigation.from_dict(
+        {
+            "url": url,
+            **subcategories,
+            **nextpage,
+            **items,
+        }
+    )
+    requests = list(spider.parse_navigation(response, navigation))
+    urls = {request.url for request in requests}
+    assert urls == {*item_urls, nextpage_url}
+    for request in requests:
+        if request.url in item_urls:
+            assert request.callback == spider.parse_product
+        else:
+            assert request.callback == spider.parse_navigation
 
 
 @pytest.mark.parametrize(
@@ -292,7 +323,7 @@ def test_metadata():
                     "title": "Crawl strategy",
                     "description": "Determines how the start URL and follow-up URLs are crawled.",
                     "type": "string",
-                    "enum": ["navigation", "full"],
+                    "enum": ["full", "navigation", "pagination_only"],
                     "enumMeta": {
                         "full": {
                             "description": "Follow most links within the domain of URL in an attempt to discover and extract as many products as possible.",
@@ -302,12 +333,23 @@ def test_metadata():
                             "description": "Follow pagination, subcategories, and product detail pages.",
                             "title": "Navigation",
                         },
+                        "pagination_only": {
+                            "description": (
+                                "Follow pagination and product detail pages. SubCategory links are ignored. "
+                                "Use this when some subCategory links are misidentified by ML-extraction."
+                            ),
+                            "title": "Pagination Only",
+                        },
                     },
                 },
                 "extract_from": {
                     "anyOf": [{"type": "string"}, {"type": "null"}],
                     "default": None,
                     "title": "Extraction source",
+                    "description": (
+                        "Whether to perform extraction using a browser request "
+                        "(browserHtml) or an HTTP request (httpResponseBody)."
+                    ),
                     "enum": ["httpResponseBody", "browserHtml"],
                     "enumMeta": {
                         "httpResponseBody": {
@@ -343,7 +385,13 @@ def test_metadata():
                     "anyOf": [{"type": "integer"}, {"type": "null"}],
                     "default": None,
                     "title": "Max Requests",
-                    "description": "The max number of Zyte API requests allowed for the crawl.",
+                    "description": (
+                        "The maximum number of Zyte API requests allowed for the crawl.\n"
+                        "\n"
+                        "Requests with error responses that cannot be retried or exceed "
+                        "their retry limit also count here, but they incur in no costs "
+                        "and do not increase the request count in Scrapy Cloud."
+                    ),
                     "widget": "request-limit",
                 },
                 "url": {
@@ -408,3 +456,16 @@ def test_metadata():
 def test_validation_url(url, valid):
     url_re = BaseSpiderParams.model_fields["url"].metadata[0].pattern
     assert bool(re.match(url_re, url)) == valid
+
+
+def test_get_parse_product_request():
+    base_kwargs = {
+        "url": "https://example.com",
+    }
+    crawler = get_crawler()
+
+    # Crawls products outside of domains by default
+    spider = EcommerceSpider.from_crawler(crawler, **base_kwargs)
+    request = ProbabilityRequest(url="https://example.com")
+    scrapy_request = spider.get_parse_product_request(request)
+    assert scrapy_request.meta.get("allow_offsite") is True