+++ zyte_spider_templates/documentation.py
@@ -1,0 +1,6 @@
+try:
+    from enum_tools.documentation import document_enum
+except ImportError:
+
+    def document_enum(func):
+        return func